pushing coffee from DB when its selected

+++ app/src/main/java/com/cso/coffeexp/data/datasource/local/CoffeeLocalDataSource.kt
@@ -5,6 +5,6 @@ import com.cso.coffeexp.data.model.local.CoffeeEntity
 interface CoffeeLocalDataSource {
     suspend fun insertCoffee(coffeeEntity: CoffeeEntity)
     suspend fun getAllCoffees(): List<CoffeeEntity>
-    suspend fun getCoffeeById(id: Int): CoffeeEntity?
+    suspend fun getCoffeeById(id: Long): CoffeeEntity?
 
 }
